peripherals/offset_dac: Rewrite/Simplify and use LiteX's SPIMaster for the DAC SPI Xfers.
Change behaviour a little bit:

In the initial implementation (and probably also with official firmware),
the SPI DAC is run continously and Mux update is done during the 6 unused
SPI Xfer bits.

To simplify the code, we no longer run the SPI DAC continuously but just run
a simple sequence continously:
- SPI DAC Xfer for channelX.
- Mux update for channelX.
- Increment channel.
- ...

The update of the 4 offsets is then a bit slower we don't require performance
here.